Change viewer action from ips to view
