--- static/ips/assets/css/custom.css
@@ -1,120 +1,0 @@
-:root {
-  --body-bg-color: #f7f3f3;
-  --primary-color: #325c33;
-}
-
-div.container-fluid.main {
-  min-height: 100%;
-  margin-right: auto;
-  margin-left: auto;
-  max-width: 800px;
-  display: flex;
-  flex-direction: column;
-}
-
-#header {
-  position: fixed;
-  z-index: 10000;
-  left: 0;
-  top: 0;
-  width: 100%;
-  background-color: var(--primary-color);
-  color: #fff;
-  line-height: 3em;
-  box-shadow: 0 0 0.15em 0 rgba(0, 0, 0, 0.1);
-  margin-bottom: 1em;
-  display: flex;
-  align-items: center;
-  height: 4em;
-}
-
-#header img {
-  height: 2em;
-  padding-left: 1em;
-}
-
-.card {
-	margin-bottom: 20px;
-  background-color: var(--body-bg-color);
-}
-
-.icon-action {
-	margin-top: 5px;
-	float: right;
-	font-size: 80%;
-}
-
-.list-group-item .title {
-	margin-top: 5px;
-	margin-bottom: 12px;
-	font-weight: 600;
-}
-
-.textBody table {
-	table-layout: fixed;
-	width: 100%;
-
-}
-
-.buttonSpacer {
-  padding-left: 20px;
-  padding-top: 20px;
-}
-
-.tdLeft {
-  text-align: left;
-}
-
-.tdCenter {
-  text-align: center;
-}
-
-.message {
-  color: red;
-  font-size: 1.0rem;
-}
-
-.loader {
-  width: 100%;
-  height: 150px;
-  margin: 40px;
-  display: block;
-  position: relative;
-  background: #FFF;
-  box-sizing: border-box;
-}
-.loader::after {
-  content: '';  
-  width: calc(100% - 30px);
-  height: calc(100% - 30px);
-  top: 15px;
-  left: 15px;
-  position: absolute;
-  background-image: linear-gradient(100deg, transparent, rgba(255, 255, 255, 0.5) 50%, transparent 80%),
-	linear-gradient(#DDD 56px, transparent 0), /* box 1 */
-	linear-gradient(#DDD 24px, transparent 0), /* box 2 */
-	linear-gradient(#DDD 18px, transparent 0), /* box 3 */
-	linear-gradient(#DDD 66px, transparent 0); /* box 4 */
-  background-repeat: no-repeat;
-  background-size: 75px 130px, /* wave */
-					55px 56px, /* box 1 */
-					160px 30px, /* box 2 */
-					220px 20px, /* box 3 */
-					290px 56px; /* box 4 */
-  background-position: 0% 0, /* box 1 */
-						0px 0px, /* box 1 */
-						70px 5px, /* box 1 */
-						70px 38px, /* box 1 */
-						0px 66px; /* box 1 */
-  box-sizing: border-box;
-  animation: animloader 1s linear infinite;
-}
-
-@keyframes animloader {
-  0% {
-    background-position: 0% 0, 0 0, 70px 5px, 70px 38px, 0px 66px;
-  }
-  100% {
-    background-position: 150% 0, 0 0, 70px 5px, 70px 38px, 0px 66px;
-  }
-}
