+++ static/view/assets/js/config.js
