+++ static/view/assets/samples/sampleBundle.json
