+++ static/view/assets/templates/AdvanceDirectives.html
@@ -29,7 +29,7 @@
           {{each(options.ad.documentReference)}}
             {{each(@this.content)}}
               {{if (@this.attachment.contentType === "application/pdf")}}
-                <iframe id="ad-pdf-viewer" src="/ips/assets/js/pdf-js/web/viewer.html" style="position: relative;   top: 0;  bottom: 0; left: 0;   width: 100%;   height: 700px;  border: 0"></iframe>
+                <iframe id="ad-pdf-viewer" src="/view/assets/js/pdf-js/web/viewer.html" style="position: relative;   top: 0;  bottom: 0; left: 0;   width: 100%;   height: 700px;  border: 0"></iframe>
               {{#else}}
                 {{if (@this.attachment.contentType === "text/html")}}
                   <div class="ad-html-viewer" style="position: relative; top: 0; bottom: 0; left: 0; width: 100%; border: 0"></div>
@@ -70,7 +70,7 @@
     {{each(options.ad.documentReference)}}
       {{each(@this.content)}}
         {{if (@this.attachment.contentType === "application/pdf")}}
-          <iframe id="ad-pdf-viewer" src="/ips/assets/js/pdf-js/web/viewer.html" style="position: relative;   top: 0;  bottom: 0; left: 0;   width: 100%;   height: 700px;  border: 0"></iframe>
+          <iframe id="ad-pdf-viewer" src="/view/assets/js/pdf-js/web/viewer.html" style="position: relative;   top: 0;  bottom: 0; left: 0;   width: 100%;   height: 700px;  border: 0"></iframe>
         {{#else}}
           {{if (@this.attachment.contentType === "text/html")}}
             <div class="ad-html-viewer" style="position: relative; top: 0; bottom: 0; left: 0; width: 100%; border: 0"></div>
