+++ static/view/assets/templates/Allergies.html
