+++ static/view/assets/templates/Observations.html
